refactor: rename "interfaces" package into "utils"
